found (stupid) mistake ... added item-selection

+++ api/src/main/java/com/brecucoder/ordero/api/item/ItemDto.java
@@ -3,17 +3,17 @@ package com.brecucoder.ordero.api.item;
 import java.math.BigDecimal;
 
 public class ItemDto {
-    private int id;
+    private Integer id;
     private String name;
     private String description;
-    private BigDecimal price;
+    private double price;
     private int stock;
 
-    public int getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -33,11 +33,11 @@ public class ItemDto {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
